allows credentials in local scheduler. Updated some test to reflect changes. Fixing #219
